refactor: fully migrate amd64 suite
