+++ test/scripts/breakpoint_test.py
@@ -83,6 +83,45 @@ class BreakpointTest(TestCase):
         d.kill()
         d.terminate()
 
+    def test_bps_waiting(self):
+        d = debugger(RESOLVE_EXE("breakpoint_test"), auto_interrupt_on_command=True)
+
+        d.run()
+
+        bp1 = d.breakpoint("random_function")
+        bp2 = d.breakpoint(TEST_BPS_ADDRESS_1)
+        bp3 = d.breakpoint(TEST_BPS_ADDRESS_2)
+
+        counter = 1
+
+        d.cont()
+
+        while True:
+            d.wait()
+            if d.instruction_pointer == bp1.address:
+                self.assertTrue(bp1.hit_count == 1)
+                self.assertTrue(bp1.hit_on(d))
+                self.assertFalse(bp2.hit_on(d))
+                self.assertFalse(bp3.hit_on(d))
+            elif d.instruction_pointer == bp2.address:
+                self.assertTrue(bp2.hit_count == counter)
+                self.assertTrue(bp2.hit_on(d))
+                self.assertFalse(bp1.hit_on(d))
+                self.assertFalse(bp3.hit_on(d))
+                counter += 1
+            elif d.instruction_pointer == bp3.address:
+                self.assertTrue(bp3.hit_count == 1)
+                CHECK_REGISTERS(self, d)
+                self.assertTrue(bp3.hit_on(d))
+                self.assertFalse(bp1.hit_on(d))
+                self.assertFalse(bp2.hit_on(d))
+                break
+
+            d.cont()
+
+        d.kill()
+        d.terminate()
+
     def test_bp_disable(self):
         d = debugger(RESOLVE_EXE("breakpoint_test"))
