+++ test/scripts/vmwhere1_test.py
@@ -9,21 +9,21 @@
 #
 
 import string
-import unittest
+from unittest import TestCase, skipUnless
+from utils.binary_utils import RESOLVE_EXE
 
 from libdebug import debugger
+from libdebug.utils.libcontext import libcontext
 
 
-class Vmwhere1(unittest.TestCase):
-    def setUp(self):
-        pass
-
+class Vmwhere1Test(TestCase):
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
     def test_vmwhere1(self):
         flag = b""
         counter = 3
         stop = False
 
-        d = debugger(["CTF/vmwhere1", "CTF/vmwhere1_program"])
+        d = debugger([RESOLVE_EXE("CTF/vmwhere1"), RESOLVE_EXE("CTF/vmwhere1_program")])
 
         while not stop:
             for el in string.printable:
@@ -56,16 +56,19 @@ class Vmwhere1(unittest.TestCase):
 
                 d.kill()
 
+        d.terminate()
+
         self.assertEqual(
             flag, b"uiuctf{ar3_y0u_4_r3al_vm_wh3r3_(gpt_g3n3r4t3d_th1s_f14g)}"
         )
 
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
     def test_vmwhere1_callback(self):
         flag = b""
         counter = 3
         stop = False
 
-        d = debugger(["CTF/vmwhere1", "CTF/vmwhere1_program"])
+        d = debugger([RESOLVE_EXE("CTF/vmwhere1"), RESOLVE_EXE("CTF/vmwhere1_program")])
 
         def callback(d, bp):
             pass
@@ -98,10 +101,8 @@ class Vmwhere1(unittest.TestCase):
 
                 d.kill()
 
+        d.terminate()
+
         self.assertEqual(
             flag, b"uiuctf{ar3_y0u_4_r3al_vm_wh3r3_(gpt_g3n3r4t3d_th1s_f14g)}"
         )
-
-
-if __name__ == "__main__":
-    unittest.main()
