+++ test/scripts/watchpoint_test.py
@@ -4,14 +4,17 @@
 # Licensed under the MIT license. See LICENSE file in the project root for details.
 #
 
-import unittest
+from unittest import TestCase, skipUnless
+from utils.binary_utils import RESOLVE_EXE
 
 from libdebug import debugger
+from libdebug.utils.libcontext import libcontext
 
 
-class WatchpointTest(unittest.TestCase):
-    def test_watchpoint(self):
-        d = debugger("binaries/watchpoint_test", auto_interrupt_on_command=False)
+class WatchpointTest(TestCase):
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_amd64(self):
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
 
         d.run()
 
@@ -21,14 +24,14 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401111)  # mov byte ptr [global_char], 0x1
+        self.assertEqual(d.instruction_pointer, 0x401111)  # mov byte ptr [global_char], 0x1
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 0)
         self.assertEqual(wp_long.hit_count, 0)
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401124)  # mov dword ptr [global_int], 0x4050607
+        self.assertEqual(d.instruction_pointer, 0x401124)  # mov dword ptr [global_int], 0x4050607
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 0)
@@ -36,7 +39,7 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         self.assertEqual(
-            d.regs.rip, 0x401135
+            d.instruction_pointer, 0x401135
         )  # mov qword ptr [global_long], 0x8090a0b0c0d0e0f
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
@@ -44,14 +47,14 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401155)  # movzx eax, byte ptr [global_char]
+        self.assertEqual(d.instruction_pointer, 0x401155)  # movzx eax, byte ptr [global_char]
         self.assertEqual(wp_char.hit_count, 2)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 1)
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401173)  # mov rax, qword ptr [global_long]
+        self.assertEqual(d.instruction_pointer, 0x401173)  # mov rax, qword ptr [global_long]
         self.assertEqual(wp_char.hit_count, 2)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 2)
@@ -59,8 +62,10 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         d.kill()
+        d.terminate()
 
-    def test_watchpoint_callback(self):
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_callback_amd64(self):
         global_char_ip = []
         global_int_ip = []
         global_long_ip = []
@@ -68,19 +73,19 @@ class WatchpointTest(unittest.TestCase):
         def watchpoint_global_char(t, b):
             nonlocal global_char_ip
 
-            global_char_ip.append(t.regs.rip)
+            global_char_ip.append(t.instruction_pointer)
 
         def watchpoint_global_int(t, b):
             nonlocal global_int_ip
 
-            global_int_ip.append(t.regs.rip)
+            global_int_ip.append(t.instruction_pointer)
 
         def watchpoint_global_long(t, b):
             nonlocal global_long_ip
 
-            global_long_ip.append(t.regs.rip)
+            global_long_ip.append(t.instruction_pointer)
 
-        d = debugger("binaries/watchpoint_test", auto_interrupt_on_command=False)
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
 
         d.run()
 
@@ -109,6 +114,7 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         d.kill()
+        d.terminate()
 
         self.assertEqual(global_char_ip[0], 0x401111)  # mov byte ptr [global_char], 0x1
         self.assertEqual(
@@ -136,8 +142,9 @@ class WatchpointTest(unittest.TestCase):
         # There is one extra hit performed by the exit routine of libc
         self.assertEqual(wp3.hit_count, 3)
 
-    def test_watchpoint_disable(self):
-        d = debugger("binaries/watchpoint_test", auto_interrupt_on_command=False)
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_disable_amd64(self):
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
 
         d.run()
 
@@ -147,14 +154,14 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401111)  # mov byte ptr [global_char], 0x1
+        self.assertEqual(d.instruction_pointer, 0x401111)  # mov byte ptr [global_char], 0x1
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 0)
         self.assertEqual(wp_long.hit_count, 0)
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401124)  # mov dword ptr [global_int], 0x4050607
+        self.assertEqual(d.instruction_pointer, 0x401124)  # mov dword ptr [global_int], 0x4050607
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 0)
@@ -162,7 +169,7 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         self.assertEqual(
-            d.regs.rip, 0x401135
+            d.instruction_pointer, 0x401135
         )  # mov qword ptr [global_long], 0x8090a0b0c0d0e0f
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
@@ -173,7 +180,7 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401173)  # mov rax, qword ptr [global_long]
+        self.assertEqual(d.instruction_pointer, 0x401173)  # mov rax, qword ptr [global_long]
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 2)
@@ -181,9 +188,11 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         d.kill()
+        d.terminate()
 
-    def test_watchpoint_disable_reenable(self):
-        d = debugger("binaries/watchpoint_test", auto_interrupt_on_command=False)
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_disable_reenable_amd64(self):
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
 
         d.run()
 
@@ -193,14 +202,14 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401111)  # mov byte ptr [global_char], 0x1
+        self.assertEqual(d.instruction_pointer, 0x401111)  # mov byte ptr [global_char], 0x1
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 0)
         self.assertEqual(wp_long.hit_count, 0)
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401124)  # mov dword ptr [global_int], 0x4050607
+        self.assertEqual(d.instruction_pointer, 0x401124)  # mov dword ptr [global_int], 0x4050607
         self.assertEqual(wp_char.hit_count, 1)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 0)
@@ -211,7 +220,7 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
 
-        self.assertEqual(d.regs.rip, 0x401155)  # movzx eax, byte ptr [global_char]
+        self.assertEqual(d.instruction_pointer, 0x401155)  # movzx eax, byte ptr [global_char]
         self.assertEqual(wp_char.hit_count, 2)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 0)
@@ -221,7 +230,7 @@ class WatchpointTest(unittest.TestCase):
 
         d.cont()
 
-        self.assertEqual(d.regs.rip, 0x401173)  # mov rax, qword ptr [global_long]
+        self.assertEqual(d.instruction_pointer, 0x401173)  # mov rax, qword ptr [global_long]
         self.assertEqual(wp_char.hit_count, 2)
         self.assertEqual(wp_int.hit_count, 1)
         self.assertEqual(wp_long.hit_count, 1)
@@ -229,3 +238,91 @@ class WatchpointTest(unittest.TestCase):
         d.cont()
 
         d.kill()
+        d.terminate()
+
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_alias_amd64(self):
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
+
+        d.run()
+
+        d.wp("global_char", condition="rw", length=1)
+        d.watchpoint("global_int", condition="w", length=4)
+        d.watchpoint("global_long", condition="rw", length=8)
+
+        d.cont()
+
+        self.assertEqual(d.instruction_pointer, 0x401111)  # mov byte ptr [global_char], 0x1
+
+        d.cont()
+
+        self.assertEqual(d.instruction_pointer, 0x401124)  # mov dword ptr [global_int], 0x4050607
+
+        d.cont()
+
+        self.assertEqual(d.instruction_pointer, 0x401135)  # mov qword ptr [global_long], 0x8090a0b0c0d0e0f
+
+        d.cont()
+
+        self.assertEqual(d.instruction_pointer, 0x401155)  # movzx eax, byte ptr [global_char]
+
+        d.cont()
+
+        self.assertEqual(d.instruction_pointer, 0x401173)  # mov rax, qword ptr [global_long]
+
+        d.cont()
+
+        d.kill()
+        d.terminate()
+
+    @skipUnless(libcontext.platform == "amd64", "Requires amd64")
+    def test_watchpoint_callback_amd64(self):
+        global_char_ip = []
+        global_int_ip = []
+        global_long_ip = []
+
+        def watchpoint_global_char(t, b):
+            nonlocal global_char_ip
+
+            global_char_ip.append(t.instruction_pointer)
+
+        def watchpoint_global_int(t, b):
+            nonlocal global_int_ip
+
+            global_int_ip.append(t.instruction_pointer)
+
+        def watchpoint_global_long(t, b):
+            nonlocal global_long_ip
+
+            global_long_ip.append(t.instruction_pointer)
+
+        d = debugger(RESOLVE_EXE("watchpoint_test"), auto_interrupt_on_command=False)
+
+        d.run()
+
+        wp1 = d.watchpoint("global_char", condition="rw", length=1, callback=watchpoint_global_char)
+        wp2 = d.wp("global_int", condition="w", length=4, callback=watchpoint_global_int)
+        wp3 = d.watchpoint("global_long", condition="rw", length=8, callback=watchpoint_global_long)
+
+        d.cont()
+
+        d.kill()
+        d.terminate()
+
+        self.assertEqual(global_char_ip[0], 0x401111)  # mov byte ptr [global_char], 0x1
+        self.assertEqual(global_int_ip[0], 0x401124)  # mov dword ptr [global_int], 0x4050607
+        self.assertEqual(global_long_ip[0], 0x401135)  # mov qword ptr [global_long], 0x8090a0b0c0d0e0f
+        self.assertEqual(global_char_ip[1], 0x401155)  # movzx eax, byte ptr [global_char]
+        self.assertEqual(global_long_ip[1], 0x401173)  # mov rax, qword ptr [global_long]
+
+        self.assertEqual(len(global_char_ip), 2)
+        self.assertEqual(len(global_int_ip), 1)
+
+        # There is one extra hit performed by the exit routine of libc
+        self.assertEqual(len(global_long_ip), 3)
+
+        self.assertEqual(wp1.hit_count, 2)
+        self.assertEqual(wp2.hit_count, 1)
+
+        # There is one extra hit performed by the exit routine of libc
+        self.assertEqual(wp3.hit_count, 3)
